Factor exceptions out into own module
