+++ tests/test_header.py
@@ -11,13 +11,9 @@ from inspect import cleandoc
 import pytest
 
 from reuse import ReuseInfo
-from reuse.comment import CommentCreateError, CppCommentStyle
-from reuse.header import (
-    MissingReuseInfo,
-    add_new_header,
-    create_header,
-    find_and_replace_header,
-)
+from reuse.comment import CppCommentStyle
+from reuse.exceptions import CommentCreateError, MissingReuseInfo
+from reuse.header import add_new_header, create_header, find_and_replace_header
 
 # REUSE-IgnoreStart
